requirements.txt
Django==5.1.0
selenium==4.27.1
Faker==33.1.0
factory-boy==3.3.1
model-bakery==1.20.0


